Created Java Files for Todos
Created a database top retrieve information from the json file of todos. Created a todo class that gives the structure for each todo allowing us to call on and filter by different parameters. Also created the TodoController class similar to the UserController class which will allow us to manage the requests for info about todos. We will be implementing the methods of getting specific todos in this file and implementing the filter methods in the TodoDatabase. Had to comment out methods in the todo files for Gradle tests to run successfully as they are not yet implemented correctly.