Made Edits on New Java Files
Made edits in the new Todo class, TodoController class, and TodoDatabase class so that the transition from the copied user code made everything pertain to todos rather that to users, thus making it its own function database and controller for the todos.json file. Edits mainly consisted of changing words from user/users to todo/todos and aligning the new database with the new controller.